# make locks fast