--- a/doc/src/base/parallel.md
+++ b/doc/src/base/parallel.md
@@ -49,7 +49,7 @@ Base.lock
 Base.unlock
 Base.trylock
 Base.islocked
-Base.Lock
+Base.ReentrantLock
 ```
 
 ## Channels