… base/locks-mt.jl
… doc/src/base/multi-threading.md
… test/threads.jl
… test/threads_exec.jl