Now that we have the ability to express atomics (particularly the write barrier of the Task object), it is time for us to speed up and simplify our locks.

This PR does need someone to verify the synchronizations are correct, but I think this version is actually quite simple. There is a single byte that represents the state of the lock:
 - 0x0 : unlocked
 - 0x1 : locked (without other waiters)
 - 0x2 : locked with waiters

And then when we have the locked bit set, we can manipulate any of the other data (e.g. owner and reentrancy) without needing any atomic work, and if we don't hold that bit, we should not write to any other field.

after:
```
julia> using BenchmarkTools

julia> const lk = ReentrantLock()
Base.Lock(nothing, 0x00000000, 0x00, Base.GenericCondition{Base.Threads.SpinLock}(Base.InvasiveLinkedList{Task}(nothing, nothing), Base.Threads.SpinLock(0)), (0, 0, 0))

julia> const lk2 = Base.ThreadSynchronizer()
Base.GenericCondition{Base.Threads.SpinLock}(Base.InvasiveLinkedList{Task}(nothing, nothing), Base.Threads.SpinLock(0))

julia> @btime (lock(lk); unlock(lk))
  24.414 ns (0 allocations: 0 bytes)

julia> lock(lk); @btime (lock(lk); unlock(lk)); unlock(lk)
  5.378 ns (0 allocations: 0 bytes)

julia> @btime (lock(lk2); unlock(lk2))
  21.072 ns (0 allocations: 0 bytes)
```

before:
```
julia> const lk = ReentrantLock()
ReentrantLock(nothing, Base.GenericCondition{Base.Threads.SpinLock}(Base.InvasiveLinkedList{Task}(nothing, nothing), Base.Threads.SpinLock(0)), 0)

julia> const lk2 = Base.ThreadSynchronizer()
Base.GenericCondition{Base.Threads.SpinLock}(Base.InvasiveLinkedList{Task}(nothing, nothing), Base.Threads.SpinLock(0))

julia> @btime (lock(lk); unlock(lk))
  56.492 ns (0 allocations: 0 bytes)

julia> lock(lk); @btime (lock(lk); unlock(lk)); unlock(lk)
  12.392 ns (0 allocations: 0 bytes)

julia> @btime (lock(lk2); unlock(lk2))
  20.071 ns (0 allocations: 0 bytes)
```

and when being used in a larger context:
```
# with lock=false (baseline)
julia> @time open("../julia-1.6.2-linux-x86_64.tar.gz", lock=false) do f; lock(f.lock); try; while true; read(f, UInt8); end; catch; end; unlock(f.lock); end;
  0.881976 seconds (7.82 k allocations: 433.968 KiB, 1.81% compilation time)

# with lock=true (baseline)
julia> @time open("../julia-1.6.2-linux-x86_64.tar.gz", lock=true) do f; lock(f.lock); while !eof(f); read(f, UInt8); end; unlock(f.lock); end;
  3.822013 seconds (8.04 k allocations: 433.238 KiB, 0.56% compilation time)

# with lock=true (WIP)
julia> @time open("../julia-1.6.2-linux-x86_64.tar.gz", lock=true) do f; lock(f.lock); try; while true; read(f, UInt8); end; catch; end; unlock(f.lock); end;
  1.296815 seconds (7.81 k allocations: 433.577 KiB, 1.52% compilation time)

# with lock option deleted
julia> @time open("../julia-1.6.2-linux-x86_64.tar.gz") do f; lock(f.lock); try; while true; read(f, UInt8); end; catch; end; unlock(f.lock); end;
  1.060876 seconds (7.81 k allocations: 433.578 KiB, 1.97% compilation time)

note: don’t ask me why println(xs::String) got recompiled every time I ran that, as I don’t know, and I am not sure I want to know

and for completeness:

# baseline
julia> @time open("../julia-1.6.2-linux-x86_64.tar.gz", lock=true) do f; try; while true; read(f, UInt8); end; catch; end; end;
  7.587580 seconds (5.74 k allocations: 318.546 KiB, 0.25% compilation time)

# WIP
julia> @time open("../julia-1.6.2-linux-x86_64.tar.gz", lock=true) do f; try; while true; read(f, UInt8); end; catch; end; end;
  3.314802 seconds (4.35 k allocations: 255.397 KiB, 0.35% compilation time)

and for reference:

julia> @time open("../julia-1.6.2-linux-x86_64.tar.gz", lock=true) do f; read(f); end;
  0.077435 seconds (4.19 k allocations: 107.953 MiB, 9.44% compilation time)
```